v0.7

+++ RSA/rsa_encrypt_decrypt.h
@@ -1,12 +1,11 @@
 #ifndef RSA_ENCRYPT_DECRYPT_H
 #define RSA_ENCRYPT_DECRYPT_H
-
-
+#include <QtGlobal>
 class RSA_Encrypt_Decrypt
 {
 public:
     RSA_Encrypt_Decrypt();
-    unsigned long int EDcrypt(int ch, int n, int e);
+    quint64 EDcrypt(quint64 ch, quint64 n, quint64 e);
 };
 
 #endif // RSA_ENCRYPT_DECRYPT_H
